+++ src/psygnal/_evented_model_v2.py
@@ -6,17 +6,15 @@ from typing import (
     Dict,
     Iterator,
     Set,
-    Type,
-    Union,
     cast,
 )
 
-from pydantic import BaseModel, PrivateAttr
+from pydantic import PrivateAttr
 from pydantic.fields import Field, FieldInfo
 
 from ._evented_model_shared import (
     NULL,
-    EventedMetaclass,
+    _EBase,
     dataclass_transform,
 )
 from ._group import SignalGroup
@@ -28,7 +26,7 @@ if TYPE_CHECKING:
 
 
 @dataclass_transform(kw_only_default=True, field_specifiers=(Field, FieldInfo))
-class EventedModel(BaseModel, metaclass=EventedMetaclass):
+class EventedModel(_EBase):
     """A pydantic BaseModel that emits a signal whenever a field value is changed.
 
     !!! important
@@ -110,18 +108,8 @@ class EventedModel(BaseModel, metaclass=EventedMetaclass):
     # when field is changed, an event for dependent properties will be emitted.
     __field_dependents__: ClassVar[Dict[str, Set[str]]]
     __eq_operators__: ClassVar[Dict[str, "EqOperator"]]
-    __slots__ = {"__weakref__"}
-    __signal_group__: ClassVar[Type[SignalGroup]]
     # pydantic BaseModel configuration.  see:
     # https://pydantic-docs.helpmanual.io/usage/model_config/
-
-    def __init__(_model_self_, **data: Any) -> None:
-        super().__init__(**data)
-        Group = _model_self_.__signal_group__
-        # the type error is "cannot assign to a class variable" ...
-        # but if we don't use `ClassVar`, then the `dataclass_transform` decorator
-        # will add _events: SignalGroup to the __init__ signature, for *all* user models
-        _model_self_._events = Group(_model_self_)  # type: ignore [misc]
 
     def _super_setattr_(self, name: str, value: Any) -> None:
         # pydantic will raise a ValueError if extra fields are not allowed
@@ -185,16 +173,6 @@ class EventedModel(BaseModel, metaclass=EventedMetaclass):
                 if not _check_field_equality(type(self), dep, after_val, before_val):
                     getattr(self._events, dep).emit(after_val)
 
-    # expose the private SignalGroup publically
-    @property
-    def events(self) -> SignalGroup:
-        """Return the `SignalGroup` containing all events for this model."""
-        return self._events
-
-    @property
-    def _defaults(self) -> dict:
-        return _get_defaults(type(self))
-
     def reset(self) -> None:
         """Reset the state of the model to default values."""
         for name, value in self._defaults.items():
@@ -205,54 +183,6 @@ class EventedModel(BaseModel, metaclass=EventedMetaclass):
                 and not self.model_fields[name].frozen
             ):
                 setattr(self, name, value)
-
-    def update(self, values: Union["EventedModel", dict], recurse: bool = True) -> None:
-        """Update a model in place.
-
-        Parameters
-        ----------
-        values : Union[dict, EventedModel]
-            Values to update the model with. If an EventedModel is passed it is
-            first converted to a dictionary. The keys of this dictionary must
-            be found as attributes on the current model.
-        recurse : bool
-            If True, recursively update fields that are EventedModels.
-            Otherwise, just update the immediate fields of this EventedModel,
-            which is useful when the declared field type (e.g. ``Union``) can have
-            different realized types with different fields.
-        """
-        if isinstance(values, BaseModel):
-            values = values.model_dump()
-        if not isinstance(values, dict):  # pragma: no cover
-            raise TypeError(f"values must be a dict or BaseModel. got {type(values)}")
-
-        with self.events._psygnal_relay.paused():  # TODO: reduce?
-            for key, value in values.items():
-                field = getattr(self, key)
-                if isinstance(field, EventedModel) and recurse:
-                    field.update(value, recurse=recurse)
-                else:
-                    setattr(self, key, value)
-
-    def __eq__(self, other: Any) -> bool:
-        """Check equality with another object.
-
-        We override the pydantic approach (which just checks
-        ``self.model_dump() == other.model_dump()``) to accommodate more complicated
-        types like arrays, whose truth value is often ambiguous. ``__eq_operators__``
-        is constructed in ``EqualityMetaclass.__new__``
-        """
-        if not isinstance(other, EventedModel):
-            return self.model_dump() == other  # type: ignore
-
-        for f_name, _ in self.__eq_operators__.items():
-            if not hasattr(self, f_name) or not hasattr(other, f_name):
-                return False  # pragma: no cover
-            a = getattr(self, f_name)
-            b = getattr(other, f_name)
-            if not _check_field_equality(type(self), f_name, a, b):
-                return False
-        return True
 
     @classmethod
     @contextmanager
@@ -276,18 +206,3 @@ class EventedModel(BaseModel, metaclass=EventedMetaclass):
                 cls.model_config["use_enum_values"] = cast(bool, before)
             else:
                 cls.model_config.pop("use_enum_values")
-
-
-def _get_defaults(obj: Type[BaseModel]) -> Dict[str, Any]:
-    """Get possibly nested default values for a Model object."""
-    dflt = {}
-    for k, v in obj.model_fields.items():
-        d = v.get_default()
-        if (
-            d is None
-            and isinstance(v.annotation, type)
-            and issubclass(v.annotation, BaseModel)
-        ):
-            d = _get_defaults(v.annotation)  # pragma: no cover
-        dflt[k] = d
-    return dflt
